packages/engine/pyproject.toml
[project]
name = "phantom-engine"
version = "0.1.0"
description = "Phantom AI pipeline — code to Remotion animation"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50.0",
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    "groq>=0.13.0",
    "mistralai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "rich>=13.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
phantom = "phantom_engine.cli:main"
phantom-server = "phantom_engine.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/phantom_engine"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
